Update Zone#import to exclude SOA record
